+++ index/segment/fs/fst_terms_offsets_map_gen.go
@@ -2,9 +2,7 @@
 // Any changes will be lost if this file is regenerated.
 // see https://github.com/mauricelam/genny
 
-package fst
-
-import "github.com/m3db/m3ninx/doc"
+package fs
 
 // Copyright (c) 2018 Uber Technologies, Inc.
 //
@@ -26,23 +24,23 @@ import "github.com/m3db/m3ninx/doc"
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-// postingsOffsetsMapHash is the hash for a given map entry, this is public to support
+// fstTermsOffsetsMapHash is the hash for a given map entry, this is public to support
 // iterating over the map using a native Go for loop.
-type postingsOffsetsMapHash uint64
+type fstTermsOffsetsMapHash uint64
 
-// postingsOffsetsMapHashFn is the hash function to execute when hashing a key.
-type postingsOffsetsMapHashFn func(doc.Field) postingsOffsetsMapHash
+// fstTermsOffsetsMapHashFn is the hash function to execute when hashing a key.
+type fstTermsOffsetsMapHashFn func([]byte) fstTermsOffsetsMapHash
 
-// postingsOffsetsMapEqualsFn is the equals key function to execute when detecting equality of a key.
-type postingsOffsetsMapEqualsFn func(doc.Field, doc.Field) bool
+// fstTermsOffsetsMapEqualsFn is the equals key function to execute when detecting equality of a key.
+type fstTermsOffsetsMapEqualsFn func([]byte, []byte) bool
 
-// postingsOffsetsMapCopyFn is the copy key function to execute when copying the key.
-type postingsOffsetsMapCopyFn func(doc.Field) doc.Field
+// fstTermsOffsetsMapCopyFn is the copy key function to execute when copying the key.
+type fstTermsOffsetsMapCopyFn func([]byte) []byte
 
-// postingsOffsetsMapFinalizeFn is the finalize key function to execute when finished with a key.
-type postingsOffsetsMapFinalizeFn func(doc.Field)
+// fstTermsOffsetsMapFinalizeFn is the finalize key function to execute when finished with a key.
+type fstTermsOffsetsMapFinalizeFn func([]byte)
 
-// postingsOffsetsMap uses the genny package to provide a generic hash map that can be specialized
+// fstTermsOffsetsMap uses the genny package to provide a generic hash map that can be specialized
 // by running the following command from this root of the repository:
 // ```
 // make hashmap-gen pkg=outpkg key_type=Type value_type=Type out_dir=/tmp
@@ -56,73 +54,73 @@ type postingsOffsetsMapFinalizeFn func(doc.Field)
 // This will output to stdout the generated source file to use for your map.
 // It uses linear probing by incrementing the number of the hash created when
 // hashing the identifier if there is a collision.
-// postingsOffsetsMap is a value type and not an interface to allow for less painful
+// fstTermsOffsetsMap is a value type and not an interface to allow for less painful
 // upgrades when adding/removing methods, it is not likely to need mocking so
 // an interface would not be super useful either.
-type postingsOffsetsMap struct {
-	_postingsOffsetsMapOptions
+type fstTermsOffsetsMap struct {
+	_fstTermsOffsetsMapOptions
 
 	// lookup uses hash of the identifier for the key and the MapEntry value
 	// wraps the value type and the key (used to ensure lookup is correct
 	// when dealing with collisions), we use uint64 for the hash partially
 	// because lookups of maps with uint64 keys has a fast path for Go.
-	lookup map[postingsOffsetsMapHash]postingsOffsetsMapEntry
+	lookup map[fstTermsOffsetsMapHash]fstTermsOffsetsMapEntry
 }
 
-// _postingsOffsetsMapOptions is a set of options used when creating an identifier map, it is kept
+// _fstTermsOffsetsMapOptions is a set of options used when creating an identifier map, it is kept
 // private so that implementers of the generated map can specify their own options
 // that partially fulfill these options.
-type _postingsOffsetsMapOptions struct {
+type _fstTermsOffsetsMapOptions struct {
 	// hash is the hash function to execute when hashing a key.
-	hash postingsOffsetsMapHashFn
+	hash fstTermsOffsetsMapHashFn
 	// equals is the equals key function to execute when detecting equality.
-	equals postingsOffsetsMapEqualsFn
+	equals fstTermsOffsetsMapEqualsFn
 	// copy is the copy key function to execute when copying the key.
-	copy postingsOffsetsMapCopyFn
+	copy fstTermsOffsetsMapCopyFn
 	// finalize is the finalize key function to execute when finished with a
 	// key, this is optional to specify.
-	finalize postingsOffsetsMapFinalizeFn
+	finalize fstTermsOffsetsMapFinalizeFn
 	// initialSize is the initial size for the map, use zero to use Go's std map
 	// initial size and consequently is optional to specify.
 	initialSize int
 }
 
-// postingsOffsetsMapEntry is an entry in the map, this is public to support iterating
+// fstTermsOffsetsMapEntry is an entry in the map, this is public to support iterating
 // over the map using a native Go for loop.
-type postingsOffsetsMapEntry struct {
+type fstTermsOffsetsMapEntry struct {
 	// key is used to check equality on lookups to resolve collisions
-	key _postingsOffsetsMapKey
+	key _fstTermsOffsetsMapKey
 	// value type stored
 	value uint64
 }
 
-type _postingsOffsetsMapKey struct {
-	key      doc.Field
+type _fstTermsOffsetsMapKey struct {
+	key      []byte
 	finalize bool
 }
 
 // Key returns the map entry key.
-func (e postingsOffsetsMapEntry) Key() doc.Field {
+func (e fstTermsOffsetsMapEntry) Key() []byte {
 	return e.key.key
 }
 
 // Value returns the map entry value.
-func (e postingsOffsetsMapEntry) Value() uint64 {
+func (e fstTermsOffsetsMapEntry) Value() uint64 {
 	return e.value
 }
 
-// _postingsOffsetsMapAlloc is a non-exported function so that when generating the source code
+// _fstTermsOffsetsMapAlloc is a non-exported function so that when generating the source code
 // for the map you can supply a public constructor that sets the correct
 // hash, equals, copy, finalize options without users of the map needing to
 // implement them themselves.
-func _postingsOffsetsMapAlloc(opts _postingsOffsetsMapOptions) *postingsOffsetsMap {
-	m := &postingsOffsetsMap{_postingsOffsetsMapOptions: opts}
+func _fstTermsOffsetsMapAlloc(opts _fstTermsOffsetsMapOptions) *fstTermsOffsetsMap {
+	m := &fstTermsOffsetsMap{_fstTermsOffsetsMapOptions: opts}
 	m.Reallocate()
 	return m
 }
 
-func (m *postingsOffsetsMap) newMapKey(k doc.Field, opts _postingsOffsetsMapKeyOptions) _postingsOffsetsMapKey {
-	key := _postingsOffsetsMapKey{key: k, finalize: opts.finalizeKey}
+func (m *fstTermsOffsetsMap) newMapKey(k []byte, opts _fstTermsOffsetsMapKeyOptions) _fstTermsOffsetsMapKey {
+	key := _fstTermsOffsetsMapKey{key: k, finalize: opts.finalizeKey}
 	if !opts.copyKey {
 		return key
 	}
@@ -131,7 +129,7 @@ func (m *postingsOffsetsMap) newMapKey(k doc.Field, opts _postingsOffsetsMapKeyO
 	return key
 }
 
-func (m *postingsOffsetsMap) removeMapKey(hash postingsOffsetsMapHash, key _postingsOffsetsMapKey) {
+func (m *fstTermsOffsetsMap) removeMapKey(hash fstTermsOffsetsMapHash, key _fstTermsOffsetsMapKey) {
 	delete(m.lookup, hash)
 	if key.finalize {
 		m.finalize(key.key)
@@ -139,7 +137,7 @@ func (m *postingsOffsetsMap) removeMapKey(hash postingsOffsetsMapHash, key _post
 }
 
 // Get returns a value in the map for an identifier if found.
-func (m *postingsOffsetsMap) Get(k doc.Field) (uint64, bool) {
+func (m *fstTermsOffsetsMap) Get(k []byte) (uint64, bool) {
 	hash := m.hash(k)
 	for entry, ok := m.lookup[hash]; ok; entry, ok = m.lookup[hash] {
 		if m.equals(entry.key.key, k) {
@@ -153,39 +151,39 @@ func (m *postingsOffsetsMap) Get(k doc.Field) (uint64, bool) {
 }
 
 // Set will set the value for an identifier.
-func (m *postingsOffsetsMap) Set(k doc.Field, v uint64) {
-	m.set(k, v, _postingsOffsetsMapKeyOptions{
+func (m *fstTermsOffsetsMap) Set(k []byte, v uint64) {
+	m.set(k, v, _fstTermsOffsetsMapKeyOptions{
 		copyKey:     true,
 		finalizeKey: m.finalize != nil,
 	})
 }
 
-// postingsOffsetsMapSetUnsafeOptions is a set of options to use when setting a value with
+// fstTermsOffsetsMapSetUnsafeOptions is a set of options to use when setting a value with
 // the SetUnsafe method.
-type postingsOffsetsMapSetUnsafeOptions struct {
+type fstTermsOffsetsMapSetUnsafeOptions struct {
 	NoCopyKey     bool
 	NoFinalizeKey bool
 }
 
 // SetUnsafe will set the value for an identifier with unsafe options for how
 // the map treats the key.
-func (m *postingsOffsetsMap) SetUnsafe(k doc.Field, v uint64, opts postingsOffsetsMapSetUnsafeOptions) {
-	m.set(k, v, _postingsOffsetsMapKeyOptions{
+func (m *fstTermsOffsetsMap) SetUnsafe(k []byte, v uint64, opts fstTermsOffsetsMapSetUnsafeOptions) {
+	m.set(k, v, _fstTermsOffsetsMapKeyOptions{
 		copyKey:     !opts.NoCopyKey,
 		finalizeKey: !opts.NoFinalizeKey,
 	})
 }
 
-type _postingsOffsetsMapKeyOptions struct {
+type _fstTermsOffsetsMapKeyOptions struct {
 	copyKey     bool
 	finalizeKey bool
 }
 
-func (m *postingsOffsetsMap) set(k doc.Field, v uint64, opts _postingsOffsetsMapKeyOptions) {
+func (m *fstTermsOffsetsMap) set(k []byte, v uint64, opts _fstTermsOffsetsMapKeyOptions) {
 	hash := m.hash(k)
 	for entry, ok := m.lookup[hash]; ok; entry, ok = m.lookup[hash] {
 		if m.equals(entry.key.key, k) {
-			m.lookup[hash] = postingsOffsetsMapEntry{
+			m.lookup[hash] = fstTermsOffsetsMapEntry{
 				key:   entry.key,
 				value: v,
 			}
@@ -195,7 +193,7 @@ func (m *postingsOffsetsMap) set(k doc.Field, v uint64, opts _postingsOffsetsMap
 		hash++
 	}
 
-	m.lookup[hash] = postingsOffsetsMapEntry{
+	m.lookup[hash] = fstTermsOffsetsMapEntry{
 		key:   m.newMapKey(k, opts),
 		value: v,
 	}
@@ -204,24 +202,24 @@ func (m *postingsOffsetsMap) set(k doc.Field, v uint64, opts _postingsOffsetsMap
 // Iter provides the underlying map to allow for using a native Go for loop
 // to iterate the map, however callers should only ever read and not write
 // the map.
-func (m *postingsOffsetsMap) Iter() map[postingsOffsetsMapHash]postingsOffsetsMapEntry {
+func (m *fstTermsOffsetsMap) Iter() map[fstTermsOffsetsMapHash]fstTermsOffsetsMapEntry {
 	return m.lookup
 }
 
 // Len returns the number of map entries in the map.
-func (m *postingsOffsetsMap) Len() int {
+func (m *fstTermsOffsetsMap) Len() int {
 	return len(m.lookup)
 }
 
 // Contains returns true if value exists for key, false otherwise, it is
 // shorthand for a call to Get that doesn't return the value.
-func (m *postingsOffsetsMap) Contains(k doc.Field) bool {
+func (m *fstTermsOffsetsMap) Contains(k []byte) bool {
 	_, ok := m.Get(k)
 	return ok
 }
 
 // Delete will remove a value set in the map for the specified key.
-func (m *postingsOffsetsMap) Delete(k doc.Field) {
+func (m *fstTermsOffsetsMap) Delete(k []byte) {
 	hash := m.hash(k)
 	for entry, ok := m.lookup[hash]; ok; entry, ok = m.lookup[hash] {
 		if m.equals(entry.key.key, k) {
@@ -235,7 +233,7 @@ func (m *postingsOffsetsMap) Delete(k doc.Field) {
 
 // Reset will reset the map by simply deleting all keys to avoid
 // allocating a new map.
-func (m *postingsOffsetsMap) Reset() {
+func (m *fstTermsOffsetsMap) Reset() {
 	for hash, entry := range m.lookup {
 		m.removeMapKey(hash, entry.key)
 	}
@@ -244,10 +242,10 @@ func (m *postingsOffsetsMap) Reset() {
 // Reallocate will avoid deleting all keys and reallocate a new
 // map, this is useful if you believe you have a large map and
 // will not need to grow back to a similar size.
-func (m *postingsOffsetsMap) Reallocate() {
+func (m *fstTermsOffsetsMap) Reallocate() {
 	if m.initialSize > 0 {
-		m.lookup = make(map[postingsOffsetsMapHash]postingsOffsetsMapEntry, m.initialSize)
+		m.lookup = make(map[fstTermsOffsetsMapHash]fstTermsOffsetsMapEntry, m.initialSize)
 	} else {
-		m.lookup = make(map[postingsOffsetsMapHash]postingsOffsetsMapEntry)
+		m.lookup = make(map[fstTermsOffsetsMapHash]fstTermsOffsetsMapEntry)
 	}
 }
